Maps Update + Methods Update
