+++ Projects/Animal Trainer/Maps/Route1.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Animal_Trainer.Maps;
+
+class Route1 : Map
+{
+	public override char[][] SpriteSheet()
+	{
+		return new char[][]
+		{
+			"gggfgggggf  fgggggfg".ToCharArray(),
+			"gggfgggggf  fgggggfg".ToCharArray(),
+			"gggfgggggf  fgggggfg".ToCharArray(),
+			"gggfgggggf  fgggggfg".ToCharArray(),
+			"gggfgggggf  fgggggfg".ToCharArray(),
+			"gggfffffff  fffffffg".ToCharArray(),
+			"gggfgggggg  ggggggfg".ToCharArray(),
+			"gggfgggggg  ggggggfg".ToCharArray(),
+			"gggfgggggT        fg".ToCharArray(),
+			"gggfrrrrrTrrrr    fg".ToCharArray(),
+			"gggTgggggTGGGGGGGGfg".ToCharArray(),
+			"gggTgggggTGGGGGGGGfg".ToCharArray(),
+			"gggTgggggTGGGGGGGGfg".ToCharArray(),
+			"gggTrrrrrTGGGGGGGGfg".ToCharArray(),
+			"gggTgggggggg      fg".ToCharArray(),
+			"gggTgggggggg      fg".ToCharArray(),
+			"gggTgggggggg  GGGGfg".ToCharArray(),
+			"gggTTTrrrrTTTTGGGGfg".ToCharArray(),
+			"gggfgggggggg  GGGGfg".ToCharArray(),
+			"gggfgggggggg  GGGGfg".ToCharArray(),
+			"gggfgg        ggggfg".ToCharArray(),
+			"gggfgg        ggggfg".ToCharArray(),
+			"gggfgg  ggggggggggfg".ToCharArray(),
+			"gggfr rrr rrrrrrrrfg".ToCharArray(),
+			"gggfgg            fg".ToCharArray(),
+			"gggfgg            fg".ToCharArray(),
+			"gggfgg      GGGG  fg".ToCharArray(),
+			"gggfTTTTTTTTGGGGrrfg".ToCharArray(),
+			"gggfggggggggGGGG  fg".ToCharArray(),
+			"gggfggggggggGGGG  fg".ToCharArray(),
+			"gggf              fg".ToCharArray(),
+			"gggTrr   srrrrrrrrTg".ToCharArray(),
+			"gggTggGGGG  ggGGGGTg".ToCharArray(),
+			"gggTggGGGG  ggGGGGTg".ToCharArray(),
+			"gggTggGGGG  GGGGggTg".ToCharArray(),
+			"gggTGGGGgg  GGGGggTg".ToCharArray(),
+			"gggTffffffGGffffffTg".ToCharArray(),
+			"gggTgggggfGGfgggggTg".ToCharArray(),
+			"gggTgggggfGGfgggggTg".ToCharArray(),
+			"gggfgggggf00fgggggfg".ToCharArray(),
+		};
+	}
+}
